+++ scripts/serve/serve-static.js
@@ -2,6 +2,7 @@
 const express = require('express');
 const path = require('path');
 const webpack = require('webpack');
+const clone = require('clone');
 const MemoryFS = require('memory-fs');
 const mime = require('mime-types');
 
@@ -19,6 +20,9 @@ const compile = (webpackConfig, disk) => (
         console.log('[Terra-Toolkit:serve-static] Webpack compiled unsuccessfully');
         reject(err || new Error(stats.toJson().errors));
       } else {
+        if (stats.hasWarnings()) {
+          console.warn(stats.toJson().warnings);
+        }
         console.log('[Terra-Toolkit:serve-static] Webpack compiled successfully');
         resolve([webpackConfig.output.path, compiler.outputFileSystem]);
       }
@@ -34,9 +38,15 @@ const generateSite = (site, config, disk, production) => {
   }
 
   if (config) {
-    let webpackConfig = config;
-    if (typeof webpackConfig === 'function') {
-      webpackConfig = webpackConfig(undefined, { p: production });
+    let webpackConfig;
+    if (typeof config === 'function') {
+      webpackConfig = config(undefined, { p: production });
+    } else {
+      webpackConfig = clone(config);
+    }
+
+    if (!(webpackConfig.output || {}).path) {
+      webpackConfig.output = Object.assign({}, webpackConfig.output, { path: '/dist' });
     }
 
     return compile(webpackConfig, disk);
@@ -45,8 +55,25 @@ const generateSite = (site, config, disk, production) => {
   return Promise.reject(new Error('[Terra-Toolkit:serve-static] No webpack configuration provided.'));
 };
 
+// Set the test locale for the html file
+const setSiteLocale = (fileContent, locale) => {
+  const localeComment = `<!-- Terra-toolkit's serve-static set the site locale to ${locale}.-->`;
+  const content = fileContent.replace(/<html/, `${localeComment}\n<html`);
+
+  const langPattern = /lang="[a-zA-Z0-9-]*"/;
+  const langLocale = `lang="${locale}"`;
+
+  const isLanguageSet = content.match(langPattern);
+  if (isLanguageSet) {
+    return content.replace(langPattern, langLocale);
+  }
+
+  return content.replace(/<html/, `<html ${langLocale}`);
+};
+
+
 // Setup an app server that reads from a filesystem.
-const virtualApp = (site, index, fs) => {
+const virtualApp = (site, index, locale, fs, verbose) => {
   const app = express();
 
   // Setup a catch all route, we can't use 'static' because we need to use a virtual file system
@@ -61,31 +88,43 @@ const virtualApp = (site, index, fs) => {
     const filepath = `${site}${filename}`.replace(/(\?).*/, '');
 
     if (fs.existsSync(filepath)) {
-      res.setHeader('content-type', mime.contentType(path.extname(filename)));
-      res.send(fs.readFileSync(filepath));
-    } else if (filename === '/favicon.ico') {
-      res.sendStatus(200);
+      if (filename === '/favicon.ico') {
+        res.sendStatus(200);
+        return;
+      }
+
+      const fileExt = path.extname(filename);
+      res.setHeader('content-type', mime.contentType(fileExt));
+
+      let fileContent = fs.readFileSync(filepath, 'utf8');
+      if (fileExt === '.html') {
+        fileContent = setSiteLocale(fileContent, locale);
+      }
+
+      res.send(fileContent);
     } else {
       next();
     }
   });
 
-  app.use((req, res, next) => {
-    const err = new Error(`Not Found: ${req.originalUrl}`);
-    err.status = 404;
-    next(err);
-  });
+  if (verbose) {
+    app.use((req, res, next) => {
+      const err = new Error(`Not Found: ${req.originalUrl}`);
+      err.status = 404;
+      next(err);
+    });
 
-  // error handler
-  app.use((err, req, res) => {
-    // set locals, only providing error in development
-    res.locals.message = err.message;
-    res.locals.error = err;
+    // error handler
+    app.use((err, req, res) => {
+      // set locals, only providing error in development
+      res.locals.message = err.message;
+      res.locals.error = err;
 
-    // render the error page
-    res.status(err.status || 500);
-    res.render('error');
-  });
+      // render the error page
+      res.status(err.status || 500);
+      res.render('error');
+    });
+  }
 
   return Promise.resolve(app);
 };
@@ -98,9 +137,9 @@ const staticApp = (site) => {
 };
 
 // Setup an app for either a virtual site or a static site. If a file system is not provided, serve the static site.
-const serveSite = (site, fs, index) => {
+const serveSite = (site, fs, index, locale, verbose) => {
   if (fs) {
-    return virtualApp(site, index, fs);
+    return virtualApp(site, index, locale, fs, verbose);
   }
 
   return staticApp(site);
@@ -108,13 +147,17 @@ const serveSite = (site, fs, index) => {
 
 // Generate a site if not provided and spin up an express server to serve the site.
 const serve = (options) => {
-  const { site, config, port, disk, index, production, host } = options;
+  const {
+    site, config, port, disk, index, locale, production, host, verbose,
+  } = options;
+
   const appPort = port || 8080;
   const appIndex = index || 'index.html';
+  const appLocale = locale || process.env.LOCALE || 'en';
 
-  return generateSite(site, config, disk, production).then(
-    ([sitePath, fs]) => serveSite(sitePath, fs, appIndex)).then(
-    (app) => {
+  return generateSite(site, config, disk, production)
+    .then(([sitePath, fs]) => serveSite(sitePath, fs, appIndex, appLocale, verbose))
+    .then((app) => {
       const server = app.listen(appPort, host);
       console.log(`[Terra-Toolkit:serve-static] Server started listening at port:${appPort}`);
       return server;