+++ scripts/serve/serve.js
@@ -11,10 +11,12 @@ const webpackServe = (options) => {
     config,
     ...(port) && { port },
     ...(host) && { host },
-    dev: { stats: {
-      colors: true,
-      children: false,
-    } },
+    dev: {
+      stats: {
+        colors: true,
+        children: false,
+      },
+    },
   };
 
   serve(serveConfig);