+++ src/wdio/services/ServeStaticService.js
@@ -1,26 +1,30 @@
-import serve from '../../../scripts/serve/serve-static';
+import serveStatic from '../../../scripts/serve/serve-static';
 import SERVICE_DEFAULTS from '../../../config/wdio/services.default-config';
 
-const SERVE_STATIC_DEFAULTS = SERVICE_DEFAULTS.serveStatic;
+const { serveStatic: SERVE_STATIC_DEFAULTS } = SERVICE_DEFAULTS;
 
 export default class ServeStaticService {
   async onPrepare(config = {}) {
-    if (!config.webpackConfig) {
+    const webpackConfig = config.webpackConfig;
+
+    if (!webpackConfig) {
       // eslint-disable-next-line no-console
       console.warn('[Terra-Toolkit:serve-static] No webpack configuration provided');
       return;
     }
 
-    const webpackConfig = config.webpackConfig;
+    const verbose = config.logLevel !== 'silent';
     const port = (config.serveStatic || {}).port || SERVE_STATIC_DEFAULTS.port;
     const index = (config.serveStatic || {}).index || SERVE_STATIC_DEFAULTS.index;
+    // Explicitly not providing a fallback locale. Providing a fallback will lock the locale for all test runs when using the tt-wdio-runner.
+    const locale = (config || {}).locale;
 
-    // If no output is provided, define one.
-    if (!(webpackConfig.output || {}).path) {
-      webpackConfig.output = Object.assign({}, webpackConfig.output, { path: '/dist' });
+    // Ensure the server was properly shut down.
+    if (this.server) {
+      await this.stop();
     }
 
-    await ServeStaticService.startService(webpackConfig, port, index).then((server) => {
+    await ServeStaticService.startService(webpackConfig, port, index, locale, verbose).then((server) => {
       this.server = server;
     });
   }
@@ -29,8 +33,10 @@ export default class ServeStaticService {
     await this.stop();
   }
 
-  static startService(config, port, index) {
-    return serve({ config, port, index, production: true });
+  static startService(config, port, index, locale, verbose) {
+    return serveStatic({
+      config, port, index, production: true, locale, verbose,
+    });
   }
 
   stop() {