+++ tests/test.config.js
@@ -1,7 +1,6 @@
 const path = require('path');
 const HtmlWebpackPlugin = require('html-webpack-plugin');
 
-
 module.exports = {
   entry: {
     index: path.join(__dirname, 'fixtures', 'index'),
@@ -42,6 +41,11 @@ module.exports = {
       chunks: ['index'],
       filename: './inaccessible-text.html',
     }),
+    new HtmlWebpackPlugin({
+      template: path.join(__dirname, 'fixtures', 'i18n.html'),
+      chunks: ['index'],
+      filename: './i18n.html',
+    }),
   ],
-  mode: 'development',
+  mode: 'production',
 };